utils: vaapi: prepare struct SurfacePoolHandle for inter-drc
Prepare struct SurfacePoolhandle for inter-frame DRC. This makes sure
that the pool will reject stale handles, as well as remove stale handles
when the resolution changes. The end result is that we can keep using
the same pool during DRC instead of creating a new one.

This allows for two optimizations:

- To make sure there is only one pool active at a time.
- To reuse surfaces during DRC if they're large enough, reducing the
  number of costly allocations.

These optimizations will follow in future commits.