utils: vaapi: reuse context and/or surfaces when applicable
When new stream parameters are found, some codecs can reuse the old
context if only the resolution changes. This is notably the case in VP9
and AV1.

We can also reuse the old surfaces if they fit, specially since we are
letting the driver choose their internal (tiled) format, mapping
on-the-fly to the format chosen by the client.

Implement these changes by stealing the resources from the old state
where applicable.